+++ projects/rocprofiler-compute/src/argparser.py
@@ -633,11 +633,29 @@ Examples:
         help="\t\tMode of spatial multiplexing.",
     )
     analyze_group.add_argument(
-        "-o",
-        "--output",
+        "--output-format",
         metavar="",
-        dest="output_file",
-        help="\t\tSpecify an output file to save analysis results.",
+        dest="output_format",
+        choices=["stdout", "txt", "csv", "db"],
+        default="stdout",
+        help=(
+            "\t\tSet the format of output file or folder containing analysis data.\n"
+            "\t\tBy default, file or folder created will "
+            "have the name rocprof_compute_<uuid>.\n"
+            "\t\tFile or folder name can be overriden using --output-name.\n"
+            "\t\tDefault output format is stdout which will not "
+            "generate any file/folder.\n"
+        ),
+    )
+    analyze_group.add_argument(
+        "--output-name",
+        metavar="",
+        dest="output_name",
+        help=(
+            "\t\tOverride the default output file name rocprof_compue_<uuid> "
+            "with the specified name.\n"
+            "\t\tThis is only applicable when --output-format txt/csv/db is used.\n"
+        ),
     )
     analyze_group.add_argument(
         "--gui",
@@ -755,12 +773,6 @@ Examples:
         metavar="",
         help="\t\tSpecify the directory of customized configs.",
         default=rocprof_compute_home.joinpath("rocprof_compute_soc/analysis_configs/"),
-    )
-    analyze_advanced_group.add_argument(
-        "--save-dfs",
-        dest="df_file_dir",
-        metavar="",
-        help="\t\tSpecify the dirctory to save analysis dataframe csv files.",
     )
     analyze_advanced_group.add_argument(
         "--cols",