+++ projects/rocprofiler-compute/src/utils/utils.py
@@ -36,6 +36,7 @@ import shutil
 import subprocess
 import tempfile
 import time
+import uuid
 from pathlib import Path as path
 from typing import Optional
 
@@ -1640,3 +1641,7 @@ def parse_sets_yaml(arch):
         if set_option:
             sets_info[set_option] = set_item
     return sets_info
+
+
+def get_uuid(length=8):
+    return uuid.uuid4().hex[:length]